\name{density.gradientForest}
\alias{density.gradientForest}
\alias{density.combinedGradientForest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Gaussian kernel density estimates
}
\description{
Extracts Gaussian kernel density estimates for a specified physical variable.
}
\usage{
\method{density}{gradientForest}(x, predictor, ...)

\method{density}{combinedGradientForest}(x, predictor, gridded=F, gear, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
       an object of class \code{gradientForest} generated by \code{\link{gradientForest}} or
    an object of an object of class \code{combinedGradientForest} generated by \code{\link{combinedGradientForest}}.
}
  \item{predictor}{
name of predictor variable.
}
  \item{gridded}{
if \code{TRUE}, supply density on the predictor grid that was used to create the \code{combinedGradientForest} object.
This ensures that densities for different gears, or the combined density, are aligned with the same predictor values. Default \code{FALSE}.
}
  \item{gear}{
name of the gear for which density is required. If missing, the combined density is supplied.
}
  \item{\dots}{
further arguments passed to or from other methods.
}
}
\value{
  Provides Gaussian kernel density estimates for a specified physical variable.
}
\references{
Ellis, N., Smith, S.J., and Pitcher, C.R. (2012) Gradient Forests: calculating importance
gradients on physical predictors. \emph{Ecology}, \bold{93}, 156--168.}
\author{
N. Ellis, CSIRO, Cleveland, Australia. <Nick.Ellis@csiro.au>
}


\seealso{
\code{\link{gradientForest}}, ~~~
}
\examples{
data(CoMLsimulation)
preds <- colnames(Xsimulation)
specs <- colnames(Ysimulation)
f1 <- gradientForest(data.frame(Ysimulation,Xsimulation), preds, specs, ntree=10)
density(f1,preds[1])

f2 <- gradientForest(data.frame(Ysimulation,Xsimulation), preds, specs[1:6+6], ntree=10)
f12 <- combinedGradientForest(west=f1,east=f2)
density(f12,preds[1])

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
