\name{predict.gradientForest}
\alias{predict.gradientForest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Predict for  gradientForest
}
\description{
Predict method for gradientForest objects.
}
\usage{
\method{predict}{gradientForest}(object, newdata, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
an object of class "gradientForest" generated by \code{\link{gradientForest}} 
}
  \item{newdata}{
An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.
}
  \item{\dots}{
%%    further arguments passed to or from other methods.
}
}
\details{
Nick:  I need some details on the prediction method.
}
\value{
\code{predict.gradientForest} returns a dataframe with --- need details here.
}
\references{
Ellis, N., Smith, S.J., and C.R. Pitcher. (2010) Gradient Forests: calculating importance
gradients on physical predictors. (submitted).
}
\author{
N. Ellis, CSIRO, Cleveland, Australia. <Nick.Ellis@csiro.au>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{gradientForest}}
}
\examples{
data(CoMLsimulation)
preds <- colnames(Xsimulation)
specs <- colnames(Ysimulation)
f1 <- gradientForest(data.frame(Ysimulation,Xsimulation), preds, specs, ntree=10)
f1.pred<-predict(f1)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
