\name{print.gradientForest}
\alias{print.gradientForest}
\alias{print.combinedGradientForest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Printing \code{gradientForest} or \code{combinedGradientForest} objects
}
\description{
Print method for \code{gradientForest} and \code{combinedGradientForest} objects.}
\usage{
\method{print}{gradientForest}(x, ...)

\method{print}{combinedGradientForest}(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    an object of class \code{gradientForest} generated by \code{\link{gradientForest}} or
    an object of an object of class \code{combinedGradientForest} generated by \code{\link{combinedGradientForest}}.
  }
  \item{\dots}{
further arguments passed to or from other methods.
}
}
\value{

Provides a brief summary of of the number of trees in the forest for the number of species with positive \eqn{R^2} as well as a list of
the important predictor variables.   If \code{x} is a \code{combinedGradientForest} object, this function provides
the number of gears (surveys) that were combined. 

}
\references{
Ellis, N., Smith, S.J., and Pitcher, C.R. (2012). Gradient Forests: calculating importance
gradients on physical predictors. \emph{Ecology}, \bold{93}, 156--168.}
\author{
N. Ellis, CSIRO, Cleveland, Australia. <Nick.Ellis@csiro.au>
}

\seealso{
\code{\link{gradientForest}}, \code{\link{combinedGradientForest}}
}
\examples{
data(CoMLsimulation)
preds <- colnames(Xsimulation)
specs <- colnames(Ysimulation)
f1 <- gradientForest(data.frame(Ysimulation,Xsimulation), preds, specs, ntree=10)
f1
f2 <- gradientForest(data.frame(Ysimulation,Xsimulation), preds, specs[1:6+6], ntree=10)
f12 <- combinedGradientForest(west=f1,east=f2)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
\keyword{ print }% __ONLY ONE__ keyword per line
